+++ CHANGELOG.md
@@ -11,3 +11,8 @@ Add visual hint with a gremlin icon in the gutter.
 ## 0.3.0
 
 Change the gremlin icon
+
+## 0.5.0
+
+Now finds right double quotation marks
+PNG icon (SVG icons not allowed anymore)